Items, Inventory, and More
I have added a new class for Items, Inventory Items, and a new Player class. I moved the PlayerMovement class into the new Player class, and have created a List<InventoryItem> in Player. I have also added a method for adding Items to the player inventory, to be used in the future.