Remove InventoryItem.cs && PlayerMovement.cs
Since the entirety of PlayerMovement was added to the Player class, PlayerMovement.cs was removed. So, Player.cs can be attached to the player in place of PlayerMovement. Additionally, I would like to redo the Inventory from the beginning and make it a bit cleaner, so we're going to take InventoryItem out from now. I also removed the calls for using InventoryItem in Player, and will replace it with a more structured inventory.