Inventory beginning, GameManager, Inventory
I deleted the Player.cs class and put back in the PlayerMovement.cs class. I realized that in order for the GameManager to work, we need to create an empty object called GameManager. Then, we attach the GameManager script Connor made the other day to the GameManager object. Now, that object is persistent. Then, we attach things such as the inventory to the GameManager object, along with the GameManager script. Because of this, there is no longer a need for Player.cs, so PlayerMovement.cs will be needed.