Added functional ItemPickup
To create an item: add a GameObject to the scene. Then, drag the ItemPickup script to the object. Then, create a new Item asset in the Items folder. Add that to the item GameObject. Congrats!!! You have a new item! Now, you can walk over to the item in the scene, and right click it to pick it up.