Temporal Blink Half-Finished
Time-Blink now officially works, with a few bugs to fix:
    Throw doesn't work after Time-Blink
    Player's animation state isn't saved and returned to

The mechanic is mapped to the "A" key and now has only pushing the key to "save" and pushing it again to "load". Later in development, we may give time-blink a permanent key-binding and two keys for save and load.