Updated UI and Inventory Pickup
The canvas is now easier to understand, and the dialogue box is cleaner and actually drops off and slides on screen for dialogue. Dialogue also works now. Interfacing to an object has been hardcoded to spacebar, which will have to be replaced with an input key from the input manager later.